requirements.txt
alembic==0.9.9
astroid==1.6.1
attrs==18.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
codecov==2.0.15
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
flasgger==0.8.1
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
future==0.16.0
hashids==1.2.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.1.0
nose==1.3.7
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pycparser==2.18
pylint==1.8.2
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
waitress==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
